app/requirements.txt
azure-common==1.1.28
azure-core==1.22.1
azure-identity==1.8.0
azure-mgmt-compute==26.1.0
azure-mgmt-core==1.3.0
azure-mgmt-resource==20.1.0
azure-mgmt-storage==19.1.0
azure-storage-blob==12.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
idna==3.3
isodate==0.6.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
oauthlib==3.2.0
portalocker==2.4.0
prometheus-client==0.13.1
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.8
